Fix the bug that was causing farmers not to apply pesticide. This was caused by the xdim and ydim values rather than the xcol and ycol values being pulled from the pest array; pests were only seen by one farmer. Also, change the delay to an int instead of a double, and force people to have a delay count of at least 1 if they are using a threshold

+++ src/densities.h
@@ -5,4 +5,4 @@
 #include <stdlib.h>
 
 void pest_dense(double **pests, double ***land, double *paras, 
-                double *thresholds, double *delay, double *delay_count);
+                double *thresholds, int *delay, int *delay_count);
